Update patch set 3
Patch Set 3:

The default behavior of k8s is to restart a container if it exits. We could override this in the Helm chart that we use to deploy dh.
On the other hand, these exit()s will happen as a result of a defective chart that doesn't set all of the necessary environment variables, so it's something that should be caught early in testing.

Patch-set: 3
Label: Verified=0